Create bot.properties-example.md
A wiki version of the bot.properties-example file. Hosting an example file on the wiki could allow the file in the main repository to be changed to `bot.properties`. Future contributors would not need to worry about the file name when creating pull requests or commits. Also creates a useful reference document for other sections of the wiki.